[lib] Verify enough local disk space exists before downloading
Verify the working directory filesystem has enough space available
before downloading a Koji build, a Koji scratch build (a task), or a
URL that references a single RPM.  In cases where enough space is not
available, the user is notified how much was found and how much is
needed.  The warning also mentions the -w option reminding the user
how to specify a different working directory.

Fixes: #667
