+++ lib/meson.build
@@ -21,6 +21,7 @@ librpminspect_sources = [
     'bytes.c',
     'checksums.c',
     'copyfile.c',
+    'curl.c',
     'debug.c',
     'delta.c',
     'deprules.c',
@@ -30,6 +31,7 @@ librpminspect_sources = [
     'files.c',
     'flags.c',
     'free.c',
+    'fs.c',
     'init.c',
     'inspect.c',
     'inspect_addedfiles.c',
